+++ tests/functional/docker-compose.yml
@@ -16,7 +16,7 @@ services:
     entrypoint: >
       sh -c "pip install -r /fastapi/tests/requirements.txt
       && python3 /fastapi/tests/utils/wait_for_es.py
-      && pytest /fastapi/tests/src"
+      && python /fastapi/main.py"
 
   redis:
     image: redis:7-alpine
